pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "langchain>=1.2.12",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "networkx>=3.6.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
